+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+Features:
+
+- theme: move theme into separate directory and the possibility to change the theme
 
 ## [2.3.0] - 2022-01-15